Add IDE controller read and implement some of readGPT

+++ script/run.sh
@@ -1,9 +1,20 @@
+if [ "$2" == "release" ]; then
 qemu-system-x86_64 \
     -bios /usr/share/OVMF/x64/OVMF.4m.fd \
-    -device ide-hd,drive=mydrive,bus=ide.0 \
-    -drive id=mydrive,file=$1/image.img,format=raw,if=none \
+    -drive file="$1/image.img",format=raw,cache=writeback,snapshot=off \
     -m 128 \
     -debugcon stdio \
     -global isa-debugcon.iobase=0xe9 \
     -d int,cpu_reset \
-    -no-reboot
+    -no-reboot 
+elif [ "$2" == "debug" ]; then
+qemu-system-x86_64 \
+    -bios /usr/share/OVMF/x64/OVMF.4m.fd \
+    -drive file="$1/image.img",format=raw,cache=writeback,snapshot=off \
+    -m 128 \
+    -debugcon stdio \
+    -global isa-debugcon.iobase=0xe9 \
+    -d int,cpu_reset \
+    -no-reboot \
+    -S -s
+fi
